idek

+++ src/transformer/activations.cpp
@@ -4,8 +4,8 @@
 
 Tensor gelu(const Tensor& input) {
     Tensor result(input.getRows(), input.getCols());
-    for (int i = 0; i < input.getRows(); i++) {
-        for (int j = 0; j < input.getCols(); j++) {
+    for (size_t i = 0; i < input.getRows(); i++) {
+        for (size_t j = 0; j < input.getCols(); j++) {
             float x = input.getValue(i, j);
             float gelu_val = 0.5f * x * (1.0f + tanh(0.79788456f * (x + 0.044715f * x * x * x)));
             result.setValue(i, j, gelu_val);
@@ -30,9 +30,9 @@ Tensor dropout(const Tensor& input, float dropout_rate, bool training) {
     float scale = 1.0f / (1.0f - dropout_rate);
 
     if (input.getIs3D()) {
-        for (int b = 0; b < input.getBatchSize(); b++) {
-            for (int i = 0; i < input.getRows(); i++) {
-                for (int j = 0; j < input.getCols(); j++) {
+        for (size_t b = 0; b < input.getBatchSize(); b++) {
+            for (size_t i = 0; i < input.getRows(); i++) {
+                for (size_t j = 0; j < input.getCols(); j++) {
                     if (dis(gen) > dropout_rate) {
                         result.setValue(b, i, j, input.getValue(b, i, j) * scale);
                     } else {
@@ -42,8 +42,8 @@ Tensor dropout(const Tensor& input, float dropout_rate, bool training) {
             }
         }
     } else {
-        for (int i = 0; i < input.getRows(); i++) {
-            for (int j = 0; j < input.getCols(); j++) {
+        for (size_t i = 0; i < input.getRows(); i++) {
+            for (size_t j = 0; j < input.getCols(); j++) {
                 if (dis(gen) > dropout_rate) {
                     result.setValue(i, j, input.getValue(i, j) * scale);
                 } else {
